requirements.txt
streamlit==1.40.0
tensorflow==2.19.1
numpy==2.0.2
opencv-python-headless==4.10.0.84
Pillow==10.4.0
